major refactoring to enhance readability
AI supported

* Split routes into focused modules and add central register_routes(app) (utility, auth, config, rephrase, prompt, check, debug)
* Make Slack routes opt-in via settings.slack_enabled; return 503 until Bolt is initialized
* Add app lifespan bootstrap (startup.py) to wire DB, HTTP, rules, LT, LLM, etc.; load test rules into Redis when configured
* Add security headers and permissive CORS middleware (middleware.py)
* Strongly type and cache settings (get_settings, reset_settings_cache); add minimum_versions typing
* Centralize API version constants (CHECK_API_VERSION=2.4, REPHRASE_API_VERSION=1.0) and validators
* Replace mutable Pydantic defaults with Field(default_factory=...); make BaseRequestIn.config factory-based
* Extract rule engine helpers: pattern/phrase matching and utilities; use append_result to build results
* Extract text utilities (parse_word_type, german_lemmatization) for reuse
* Tighten Slack block builder and add Slack tests (route inclusion, 503 guard, blocks smoke test)
* Update tests to pass context to parse_term_replacements/fetch_configs_for_request
* Minor docstring and cleanup in model.py and review prompt (returns str | None, supports min_changes)
