+++ app/bolt.py
@@ -1,16 +1,16 @@
 from app.settings import Settings
 from app.models import ResultOut, Language
+from app.context import AppContext
 
-from slack_bolt.app.async_app import AsyncApp
+from slack_bolt.async_app import AsyncApp, AsyncRespond
 from slack_sdk.models.blocks import (
     SectionBlock,
     MarkdownTextObject,
 )
-from slack_bolt import Respond
 from slack_sdk.web.async_client import AsyncWebClient
 
 
-def get_bolt(settings: Settings):
+def get_bolt(settings: Settings, context: AppContext) -> AsyncApp:
     if settings.slack_bot_token and settings.slack_signing_secret:  # pragma: no cover
         bolt = AsyncApp(
             token=settings.slack_bot_token, signing_secret=settings.slack_signing_secret
@@ -24,6 +24,12 @@ def get_bolt(settings: Settings):
             ),
         )
 
+    # Inject AppContext into Slack Bolt's context for all listeners
+    @bolt.middleware
+    async def inject_app_context(context_, next):  # type: ignore[no-redef]
+        context_["app_context"] = context
+        return await next()
+
     return bolt
 
 
@@ -32,7 +38,7 @@ async def process_command_witty(
     language: Language,
     limit_reached: bool,
     results: list[ResultOut],
-    respond: Respond,
+    respond: AsyncRespond,
 ):  # pragma: no cover
     analyzed_text = f"*Analyzed*: {text}"
     if limit_reached:
@@ -73,9 +79,10 @@ async def process_command_witty(
                 )
             )
 
-            if len(result.alternatives):
+            alternatives_list = result.alternatives or []
+            if alternatives_list:
                 alternatives = ""
-                for alternative in result.alternatives:
+                for alternative in alternatives_list:
                     if alternative.remove:
                         alternatives += f"\n• ~{alternative.text}~"
                     else:
@@ -92,3 +99,4 @@ async def process_command_witty(
                 )
 
     await respond(blocks=blocks)
+    return None
