+++ app/routes/__init__.py
@@ -1,0 +1,41 @@
+"""Route registration for all API endpoints."""
+
+from app.routes import (
+    utility,
+    auth,
+    config_routes,
+    rephrase,
+    prompt,
+    check,
+    debug,
+)
+from app.settings import get_settings
+
+
+def register_routes(app):
+    # Register utility routes
+    app.include_router(utility.router, tags=["utility"])
+
+    # Register authentication routes
+    app.include_router(auth.router, tags=["auth"])
+
+    # Register configuration management routes
+    app.include_router(config_routes.router, tags=["config"])
+
+    # Register rephrase routes
+    app.include_router(rephrase.router, tags=["rephrase"])
+
+    # Register prompt routes
+    app.include_router(prompt.router, tags=["prompt"])
+
+    # Register Slack integration routes if enabled
+    if get_settings().slack_enabled:
+        from app.routes import slack as slack_module
+
+        app.include_router(slack_module.router, tags=["slack"])
+
+    # Register text checking routes (core functionality)
+    app.include_router(check.router, tags=["check"])
+
+    # Register debug routes
+    app.include_router(debug.router, tags=["debug"])
